refactor(components): update header and MobileMenu extraction
